Add files via upload
I worked on a project that involved analyzing a dataset of car listings sourced from Yad2, a popular online classifieds platform in Israel, which I obtained from Kaggle. The main goal of this project was to explore and clean the data, gain insights into the local car market, and identify key patterns within the listings. The project was implemented using Python, where I applied various data analysis techniques to process and visualize the information.